Implementation Algorithm MarkovChain and  improvements GeneratedLevel

+++ GeometryDashAI/src/com/Component/Grid.java
@@ -42,12 +42,8 @@ public class Grid extends Component {
             }
         }
     }
-
-    /**
-     * @param dt
-     */
     @Override
-    public void update(float dt) {
+    public void update(double dt) {
 
     }
