+++ GeometryDashAI/src/com/Component/TriangleBounds.java
@@ -31,13 +31,10 @@ public class TriangleBounds extends Bounds{
     {
         calculateTransform();
     }
-
-
     @Override
-    public void update(float dt) {
+    public void update(double dt) {
 
     }
-
     public static boolean checkCollision(BoxBounds b1, TriangleBounds t2)
     {
         if(t2.broadPhase(b1))
